# docs/swagger.yaml
basePath: /api
definitions:
  http.KVPair:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  http.RangeQueryResult:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
  description: Bitcask的RESTful API服务
  title: Bitcask API
  version: "1.0"
paths:
  /admin/hint:
    post:
      description: 生成hint文件，加速下次启动
      produces:
      - text/plain
      responses:
        "200":
          description: 生成hint文件成功
          schema:
            type: string
        "500":
          description: 生成hint文件失败
          schema:
            type: string
      summary: 生成hint文件
      tags:
      - admin
  /admin/merge:
    post:
      description: 合并数据文件，删除过时记录
      produces:
      - text/plain
      responses:
        "200":
          description: 合并成功
          schema:
            type: string
        "500":
          description: 合并失败
          schema:
            type: string
      summary: 执行合并操作
      tags:
      - admin
  /keys:
    get:
      description: 获取系统中所有键值对
      produces:
      - application/json
      responses:
        "200":
          description: 键值对列表
          schema:
            items:
              $ref: '#/definitions/http.KVPair'
            type: array
        "500":
          description: 扫描失败
          schema:
            type: string
      summary: 列出所有键值对
      tags:
      - keys
  /keys/{key}:
    delete:
      description: 从系统中删除指定的键值对
      parameters:
      - description: 要删除的键名
        in: path
        name: key
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: 删除成功
          schema:
            type: string
        "500":
          description: 删除失败
          schema:
            type: string
      summary: 删除指定key
      tags:
      - keys
    get:
      consumes:
      - application/json
      description: 获取存储在系统中的指定key的值
      parameters:
      - description: 查询的键名
        in: path
        name: key
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: 键值内容
          schema:
            type: string
        "404":
          description: 获取值失败
          schema:
            type: string
      summary: 获取指定key的值
      tags:
      - keys
    put:
      consumes:
      - text/plain
      description: 存储或更新键值对
      parameters:
      - description: 设置的键名
        in: path
        name: key
        required: true
        type: string
      - description: 存储的值
        in: body
        name: value
        required: true
        schema:
          type: string
      produces:
      - text/plain
      responses:
        "200":
          description: 存储成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 存储失败
          schema:
            type: string
      summary: 设置key的值
      tags:
      - keys
  /keys/range/{start}/{end}:
    get:
      description: 查询指定键范围内的键值对
      parameters:
      - description: 起始键
        in: path
        name: start
        required: true
        type: string
      - description: 结束键
        in: path
        name: end
        required: true
        type: string
      - default: 100
        description: 最大返回数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 范围内的键值对
          schema:
            items:
              $ref: '#/definitions/http.RangeQueryResult'
            type: array
        "500":
          description: 范围扫描失败
          schema:
            type: string
      summary: 范围查询键值对
      tags:
      - keys
swagger: "2.0"